separate traffic_light_utils from perception_utils
