+++ common/perception_utils/CMakeLists.txt
@@ -8,7 +8,6 @@ find_package(Boost REQUIRED)
 
 ament_auto_add_library(perception_utils SHARED
   src/predicted_path_utils.cpp
-  src/traffic_light_utils.cpp
   src/conversion.cpp
 )
