+++ common/traffic_light_utils/src/traffic_light_utils.cpp
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "perception_utils/traffic_light_utils.hpp"
+#include "traffic_light_utils/traffic_light_utils.hpp"
 
-namespace perception_utils
+namespace traffic_light_utils
 {
 
 namespace traffic_light
@@ -76,4 +76,4 @@ tf2::Vector3 getTrafficLightCenter(const lanelet::ConstLineString3d & traffic_li
 
 }  // namespace traffic_light
 
-}  // namespace perception_utils
+}  // namespace traffic_light_utils
