+++ common/traffic_light_utils/test/test_traffic_light_utils.cpp
@@ -1,0 +1,41 @@
+// Copyright 2022 TIER IV, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "traffic_light_utils/traffic_light_utils.hpp"
+
+#include <gtest/gtest.h>
+
+constexpr double epsilon = 1e-06;
+
+namespace
+{
+autoware_auto_perception_msgs::msg::ObjectClassification createObjectClassification(
+  const std::uint8_t label, const double probability)
+{
+  autoware_auto_perception_msgs::msg::ObjectClassification classification;
+  classification.label = label;
+  classification.probability = probability;
+
+  return classification;
+}
+
+}  // namespace
+
+TEST(object_classification, test_getHighestProbLabel)
+{
+  using autoware_auto_perception_msgs::msg::ObjectClassification;
+  using traffic_light_utils::getHighestProbLabel;
+
+}
+
